+++ error_handler.go
@@ -13,6 +13,30 @@ import (
 var resetColor = "\033[0m"
 var errLog = log.New(os.Stderr, "\n\n\x1b[31m", log.LstdFlags)
 
+// errorHandlers is a mapping from an error type to that error handler
+type errorHandlers map[reflect.Type]reflect.Value
+
+func (h errorHandlers) setup(handler interface{}) {
+	ht := reflect.TypeOf(handler)
+	validateErrorHandler(ht)
+
+	h[ht.In(0)] = reflect.ValueOf(handler)
+}
+
+func (h errorHandlers) copy() errorHandlers {
+	newHandlers := make(errorHandlers, len(h))
+	for typ, value := range h {
+		newHandlers[typ] = value
+	}
+	return newHandlers
+}
+
+func newErrorHandlers() errorHandlers {
+	handlers := errorHandlers{}
+	handlers.setup(DefaultErrorHandler)
+	return handlers
+}
+
 func DefaultErrorHandler(err error) (status Status, response *DefaultErrorResponse) {
 	status = 500
 	response = &DefaultErrorResponse{
@@ -48,91 +72,43 @@ func DefaultErrorHandler(err error) (status Status, response *DefaultErrorRespon
 	return
 }
 
-func WrapErrorHandler(handler interface{}) *ErrorHandlerCaller {
-	caller := &ErrorHandlerCaller{
-		originalHandler: handler,
-	}
-
-	caller.init()
-	return caller
-}
-
-type ErrorHandlerCaller struct {
-	originalHandler interface{}
-	handler         reflect.Value
-	responseType    reflect.Type
-	responseIndex   int
-	statusIndex     int
-}
-
-func (c *ErrorHandlerCaller) init() {
-	c.handler = reflect.ValueOf(c.originalHandler)
-
-	ht := reflect.TypeOf(c.originalHandler)
-	c.validate(ht)
-	c.recognizeOutParams(ht)
-}
-
-func (c *ErrorHandlerCaller) validate(ht reflect.Type) {
+func validateErrorHandler(ht reflect.Type) {
 	if ht.Kind() != reflect.Func {
 		panic(fmt.Sprintf("error handler '%s' is not a function", ht))
 	}
 
 	if ht.NumIn() != 1 {
-		panic(fmt.Sprintf("error handler '%s' must accept exactly one argument of type 'error', was '%d' arguments", ht, ht.NumIn()))
+		panic(fmt.Sprintf("error handler '%s' must accept exactly one argument implementing 'error', was '%d' arguments", ht, ht.NumIn()))
 	}
 
-	if ht.In(0) != errorInterfaceType {
-		panic(fmt.Sprintf("error handler '%s' must accept argument of type 'error' instead of type '%s'", ht, ht.In(0)))
-	}
-
-	if ht.NumOut() != 2 {
-		panic(fmt.Sprintf("error handler '%s' must return exactly two arguments, was '%d' arguments", ht, ht.NumOut()))
+	input := ht.In(0)
+	if !input.Implements(errorInterfaceType) {
+		panic(fmt.Sprintf("error handler '%s' must accept argument implementing 'error`, got type '%s'", ht, input))
 	}
 }
 
-func (c *ErrorHandlerCaller) recognizeOutParams(ht reflect.Type) {
-	errorMsg := fmt.Sprintf("error handler '%s' must return status of type 'int' and some other type as response payload; if you need an integer as payload, use 'gnext.Status' for status", ht)
-
-	switch {
-	// NOTE: order of cases below is meaningful, do not touch it without a good reason
-	case ht.Out(0) == ht.Out(1):
-		panic(errorMsg)
-	case ht.Out(0) == statusType:
-		c.statusIndex = 0
-		c.responseIndex = 1
-		c.responseType = ht.Out(1)
-	case ht.Out(1) == statusType:
-		c.statusIndex = 1
-		c.responseIndex = 0
-		c.responseType = ht.Out(0)
-	case ht.Out(0).Kind() == reflect.Int && ht.Out(1).Kind() == reflect.Int:
-		panic(errorMsg)
-	case ht.Out(0).Kind() == reflect.Int:
-		c.statusIndex = 0
-		c.responseIndex = 1
-		c.responseType = ht.Out(1)
-	case ht.Out(1).Kind() == reflect.Int:
-		c.statusIndex = 1
-		c.responseIndex = 0
-		c.responseType = ht.Out(0)
-	default:
-		panic(errorMsg)
+func newErrorHandlerCaller(handler reflect.Value) *errorHandlerCaller {
+	return &errorHandlerCaller{
+		handler:       handler,
+		defaultStatus: 500,
 	}
 }
 
-func (c *ErrorHandlerCaller) call(context *callContext) {
-	results := c.handler.Call([]reflect.Value{*context.error})
+type errorHandlerCaller struct {
+	handler       reflect.Value
+	argSetters    []argSetter
+	defaultStatus Status
+}
 
-	status := results[c.statusIndex].Convert(intType).Interface().(int)
-	response := results[c.responseIndex].Interface()
-	if status == 0 {
-		status = 500
-		response = &DefaultErrorResponse{
-			Message: "internal server error",
-			Success: false,
-		}
+func (c *errorHandlerCaller) addSetter(setter argSetter) {
+	c.argSetters = append(c.argSetters, setter)
+}
+
+func (c *errorHandlerCaller) call(ctx *callContext) {
+	results := c.handler.Call([]reflect.Value{ctx.error.Elem()})
+	ctx.status = c.defaultStatus
+
+	for i, setter := range c.argSetters {
+		setter(&results[i], ctx)
 	}
-
-	context.rawContext.JSON(status, response)
 }