+++ group.go
@@ -7,15 +7,15 @@ import (
 )
 
 type routerGroup struct {
-	pathPrefix   string
-	rawRouter    gin.IRouter
-	middlewares  []Middleware
-	Docs         *docs.Docs
-	errorHandler interface{}
+	pathPrefix    string
+	rawRouter     gin.IRouter
+	middlewares   middlewares
+	Docs          *docs.Docs
+	errorHandlers errorHandlers
 }
 
 func (g *routerGroup) OnError(errorHandler interface{}) IRoutes {
-	g.errorHandler = errorHandler
+	g.errorHandlers.setup(errorHandler)
 	return g
 }
 
@@ -55,23 +55,23 @@ func (g *routerGroup) POST(path string, handler interface{}, doc ...*docs.Endpoi
 }
 
 func (g *routerGroup) Handle(method string, path string, handler interface{}, doc ...*docs.Endpoint) IRoutes {
-	wrapper := WrapHandler(method, g.fullPath(path), g.middlewares, g.Docs, handler, g.errorHandler, doc...)
-	g.rawRouter.Handle(method, path, wrapper.getHandler())
+	wrapper := WrapHandler(method, g.fullPath(path), g.middlewares, g.Docs, handler, g.errorHandlers, doc...)
+	g.rawRouter.Handle(method, path, wrapper.requestHandler)
 	return g
 }
 
 func (g *routerGroup) Use(middleware Middleware) IRoutes {
-	g.middlewares = append(g.middlewares, middleware)
+	g.middlewares = append(g.middlewares, &middleware)
 	return g
 }
 
 func (g *routerGroup) Group(prefix string, _ ...*docs.Endpoint) IRouter {
 	return &routerGroup{
-		pathPrefix:   g.fullPath(prefix),
-		rawRouter:    g.rawRouter.Group(prefix),
-		middlewares:  append([]Middleware{}, g.middlewares...),
-		Docs:         g.Docs,
-		errorHandler: g.errorHandler,
+		pathPrefix:    g.fullPath(prefix),
+		rawRouter:     g.rawRouter.Group(prefix),
+		middlewares:   g.middlewares.copy(),
+		Docs:          g.Docs,
+		errorHandlers: g.errorHandlers.copy(),
 	}
 }